requirements.txt
allure-pytest==2.8.9
allure-python-commons==2.8.9
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
PyHamcrest==2.0.0
pyparsing==2.4.6
pytest==5.3.5
pytest-html==2.0.1
pytest-metadata==1.8.0
requests==2.22.0
selenium==3.141.0
six==1.14.0
unittest-xml-reporting==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==0.16.1
